Do "press any key to continue" on separate bat
The old version of the script only prevented closing the cmd window on
Windos whan a syntax error occurred. If a command line parsing error or
uncaught exceptin occurred it would not pause the program.

The fix for this is to move the "press any key to continue" feature to a
separate windows-only bat file and get rid of the "--quiet" flag. This
also means there is no cruft for the Linux version of the program.